Made some files for inchat, renamed some folders and made some minor changes.
