+++ src/interface_adapter/in_chat/InChatPrivateController.java
@@ -1,0 +1,4 @@
+package interface_adapter.in_chat;
+
+public class InChatPrivateController {
+}
