+++ src/interface_adapter/in_chat/InChatPrivateViewModel.java
@@ -1,0 +1,7 @@
+package interface_adapter.in_chat;
+
+import javax.swing.*;
+
+public class InChatPrivateViewModel {
+    JFrame privatechatframe = new JFrame("ngechat");
+}
