+++ src/view/contacts_list/GroupsListView.java
@@ -1,4 +1,4 @@
-package view.contacts;
+package view.contacts_list;
 
 public class GroupsListView {
 }
